add gen model

+++ md-atom/md-atom-user-core/src/main/java/com/atom/core/model/MdBrand.java
@@ -347,6 +347,6 @@ public class MdBrand {
 	/**
 	 * 
 	 */
-	private int suid;
+	private long suid;
 
 }
